+++ tests/ServiceStack.Redis.Tests/Examples/BlogPostExample.cs
@@ -15,12 +15,12 @@ namespace ServiceStack.Redis.Tests.Examples
 	{
 		public User()
 		{
-			this.BlogIds = new List<int>();
+			this.BlogIds = new List<long>();
 		}
 
-		public int Id { get; set; }
+		public long Id { get; set; }
 		public string Name { get; set; }
-		public List<int> BlogIds { get; set; }
+		public List<long> BlogIds { get; set; }
 	}
 
 	public class Blog
@@ -28,14 +28,14 @@ namespace ServiceStack.Redis.Tests.Examples
 		public Blog()
 		{
 			this.Tags = new List<string>();
-			this.BlogPostIds = new List<int>();
+			this.BlogPostIds = new List<long>();
 		}
 
-		public int Id { get; set; }
-		public int UserId { get; set; }
+		public long Id { get; set; }
+		public long UserId { get; set; }
 		public string UserName { get; set; }
 		public List<string> Tags { get; set; }
-		public List<int> BlogPostIds { get; set; }
+		public List<long> BlogPostIds { get; set; }
 	}
 
 	public class BlogPost
@@ -47,8 +47,8 @@ namespace ServiceStack.Redis.Tests.Examples
 			this.Comments = new List<BlogPostComment>();
 		}
 
-		public int Id { get; set; }
-		public int BlogId { get; set; }
+		public long Id { get; set; }
+		public long BlogId { get; set; }
 		public string Title { get; set; }
 		public string Content { get; set; }
 		public List<string> Categories { get; set; }
